+++ NEWS.md
@@ -7,6 +7,6 @@
 * Fixed index variable issue related to generating categorical data 
 * Fixed index variable issue related to generating longitudinal data
 * Fixed issue that arised When creating categorical variable in first field
-
-
-
+* Increased speed required to generate categorical data with large sample sizes
+* Categorical data can now accomodate probabilities condition on covariates
+* Fix: package data.table 1.10.0 broke genMissDataMat. genMissDataMat has been updated.